initial implementation
currently only works on NSFW channels.
there is a plan (albeit a dangerous one) to port this command to SFW environment, where it would only pull text only from SFW boards. this would still require a LOT of moderation, but with some lists and perhaps keywords it should be doable.