fixed an error with chrono branch
if the message contained a proper prefix (such as "in"), but also contained "on" even as a part of another word, the chrono branch would activate, causing unexpected results.

i.e. "remind me in 5 minutes and 10 seconds" would fire the chrono branch, taking timezones into account, which makes no sense.